Transactions Page
Different Components of transactions page is added to Routes folder including respective css files. The transactions list is just a table with one row as an example to show how it looks. Does not have to functionality of adding transactions.
The cards showing transactions also do not have any function for now.
Only ui is done.
Buttons need to have functions added to them